(VIDEO-672) Reconnect to telnyx client if requried when receiving a PN
To major changes:
- Invoke to process incoming push notification of the SDK whenever a PN is received. Inside the SDK method we are re-connecting if required.
- All the TelnyxClient delegate methods are moved to the appDelegate and the appDelegate will redirect them to the ViewContorller that implemnets VoIPDelegate
