+++ TelnyxWebRTCDemo/ViewController.swift
@@ -36,7 +36,6 @@ class ViewController: UIViewController {
         print("ViewController:: viewDidLoad()")
 
         self.telnyxClient = appDelegate.telnyxClient
-        self.telnyxClient?.delegate = self
         self.initViews()
     }
 
@@ -180,134 +179,6 @@ class ViewController: UIViewController {
     }
 }
 
-// MARK: - TxClientDelegate
-extension ViewController: TxClientDelegate {
-
-    func onSocketConnected() {
-        print("ViewController:: TxClientDelegate onSocketConnected()")
-        DispatchQueue.main.async {
-            self.socketStateLabel.text = "Connected"
-            self.connectButton.setTitle("Disconnect", for: .normal)
-        }
-        
-    }
-    
-    func onSocketDisconnected() {
-        print("ViewController:: TxClientDelegate onSocketDisconnected()")
-        DispatchQueue.main.async {
-            self.removeLoadingView()
-            self.resetCallStates()
-            self.socketStateLabel.text = "Disconnected"
-            self.connectButton.setTitle("Connect", for: .normal)
-            self.sessionIdLabel.text = "-"
-            self.settingsView.isHidden = false
-            self.callView.isHidden = true
-            self.incomingCallView.isHidden = true
-        }
-    }
-    
-    func onClientError(error: Error) {
-        print("ViewController:: TxClientDelegate onClientError() error: \(error)")
-        DispatchQueue.main.async {
-            self.removeLoadingView()
-            self.incomingCallView.isHidden = true
-            self.telnyxClient?.disconnect()
-
-            let alert = UIAlertController(title: "WebRTC error", message: error.localizedDescription, preferredStyle: .alert)
-            alert.addAction(UIAlertAction(title: "OK", style: .cancel, handler: {_ in
-                self.navigationController?.popViewController(animated: true)
-            }))
-            self.present(alert, animated: true)
-        }
-    }
-    
-    func onClientReady() {
-        print("ViewController:: TxClientDelegate onClientReady()")
-        DispatchQueue.main.async {
-            self.removeLoadingView()
-            self.socketStateLabel.text = "Client ready"
-            self.settingsView.isHidden = true
-            self.callView.isHidden = false
-            self.incomingCallView.isHidden = true
-        }
-    }
-    
-    func onSessionUpdated(sessionId: String) {
-        print("ViewController:: TxClientDelegate onSessionUpdated() sessionId: \(sessionId)")
-        DispatchQueue.main.async {
-            self.sessionIdLabel.text = sessionId
-        }
-    }
-
-    func onIncomingCall(call: Call) {
-        guard let callId = call.callInfo?.callId else {
-            print("ViewController:: TxClientDelegate onIncomingCall() Error unknown call UUID")
-            return
-        }
-        print("ViewController:: TxClientDelegate onIncomingCall() Error unknown call UUID: \(callId)")
-        
-        if let currentCallUUID = self.currentCall?.callInfo?.callId {
-            appDelegate.executeEndCallAction(uuid: currentCallUUID) //Hangup the previous call if there's one active
-        }
-        self.currentCall = call //Update the current call with the incoming call
-        self.incomingCall = true
-        DispatchQueue.main.async {
-            self.updateButtonsState()
-            self.incomingCallView.isHidden = false
-            self.callView.isHidden = true
-            //Hide the keyboard
-            self.view.endEditing(true)
-        }
-        appDelegate.newIncomingCall(from: call.callInfo?.callerName ?? "Unknown", uuid: callId)
-    }
-
-    func onRemoteCallEnded(callId: UUID) {
-        print("ViewController:: TxClientDelegate onRemoteCallEnded() callId: \(callId)")
-        let reason = CXCallEndedReason.remoteEnded
-        if let provider = appDelegate.callKitProvider {
-            provider.reportCall(with: callId, endedAt: Date(), reason: reason)
-        }
-    }
-    
-    func onRemoteCallAnswered(call: Call) {
-        self.currentCall = call
-        self.incomingCall = false
-        DispatchQueue.main.async {
-            self.updateButtonsState()
-            self.incomingCallView.isHidden = false
-            self.callView.isHidden = true
-            //Hide the keyboard
-            self.view.endEditing(true)
-        }
-    }
-
-    func onCallStateUpdated(callState: CallState, callId: UUID) {
-        DispatchQueue.main.async {
-            switch (callState) {
-            case .CONNECTING:
-                break
-            case .RINGING:
-                break
-            case .NEW:
-                break
-            case .ACTIVE:
-                self.incomingCallView.isHidden = true
-                self.callView.isHidden = false
-                break
-            case .DONE:
-                if let currentCallId = self.currentCall?.callInfo?.callId,
-                   currentCallId == callId {
-                    self.currentCall = nil // clear current call
-                }
-                self.resetCallStates()
-                break
-            case .HELD:
-                break
-            }
-            self.updateButtonsState()
-        }
-    }
-}
 // MARK: - UIIncomingCallViewDelegate
 /**
  Handle Incoming Call Views events
